Update
Delete functions now working properly. After clicking remove it shows the information for that donation. you verify by entering the donation number and then you are sent back to the list of donations with a message verifying the removal. adding a user is working properly, you add a user and you get a successful add message and redirected to the list of donations